floatをintegerに変更

+++ db/migrate/20201212055244_create_squat_weight_records.rb
@@ -1,7 +1,7 @@
 class CreateSquatWeightRecords < ActiveRecord::Migration[6.0]
   def change
     create_table :squat_weight_records do |t|
-      t.float      :squat_weight
+      t.integer      :squat_weight
       t.date       :squat_day
       t.references :user,           foreign_key: true
       t.timestamps
